feat: export more apis from server core
