+++ packages/server/core/src/base/utils/index.ts
@@ -8,3 +8,4 @@ export * from './error';
 export * from './warmup';
 export * from './entry';
 export * from './path';
+export * from './templates';
